Add: Builtin LSP semantic highlight support.
modified: README.md
renamed: colors/juliana.vim -> colors/juliana.lua
renamed:  lua/nvim-juliana/juliana.lua -> lua/nvim-juliana/init.lua
modified: lua/nvim-juliana/palette.lua
