+++ colors/juliana.lua
@@ -1,0 +1,1 @@
+require('nvim-juliana').load()
