--- colors/juliana.vim
@@ -1,1 +1,0 @@
-lua require('nvim-juliana.juliana').load()
