+++ lua/nvim-juliana/palette.lua
@@ -15,20 +15,21 @@ local st = {
     pink = '#cc8ec6',
 }
 local normal = {
-    gray0 = '#333333',
-    gray1 = '#303841',
-    gray2 = '#46525C',
-    gray3 = '#A6ACB8',
-    gray4 = '#d5dee6',
-    white = '#FFFFFF',
-    blue = '#5c99d6',
-    cyan = '#5FB4B4',
-    green = '#99C794',
-    yellow = '#FAC761',
-    orange1 = '#F9AE58',
-    orange2 = '#F97B58',
-    red = '#EC5F66',
-    pink = '#cc8ec6',
+  black = '#333333',
+  bg0 = '#272E35',
+  bg1 = '#303841',
+  gray0 = '#46525C',
+  gray1 = '#A6ACB8',
+  fg = '#d5dee6',
+  white1 = '#FFFFFF',
+  blue = '#5c99d6',
+  cyan = '#5FB4B4',
+  green = '#99C794',
+  yellow = '#FAC761',
+  orange1 = '#F9AE58',
+  orange2 = '#F97B58',
+  red = '#EC5F66',
+  pink = '#cc8ec6',
 }
 
 return {
